pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aioeloverblik"
version = "0.1.0"
description = "Async Python wrapper for Eloverblik.dk API (Customer & ThirdParty)"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
  { name = "William", email = "william@heitech.dk" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "httpx>=0.28.1",
  "pydantic>=2.12.5",
]

[project.urls]
Repository = "https://github.com/will2022/aioeloverblik"
Changelog = "https://github.com/will2022/aioeloverblik/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
  "python-dotenv>=1.2.1",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "respx>=0.22.0",
  "ruff>=0.14.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/aioeloverblik"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "COM"]
ignore = []

